refactor(types): move type definitions to separate Types.ts module
- Extract shared type definitions from `Documents.ts` to new `Types.ts` file
- Update imports in dependent modules (`SearchOnlyCollection.ts`, `Utils.ts`)
- Add deprecation notice for type imports from `Documents.ts`
- Maintain backward compatibility with re-exports
